ai-service/requirements.txt
# AI Service — Python dependencies
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Add ML/model libs here as needed, e.g.:
# transformers>=4.38.0
# torch>=2.2.0
# Pillow>=10.2.0
